Port the project to TypeScript
