+++ .eslintignore
@@ -1,0 +1,2 @@
+*.d.ts
+/index.js
