+++ .eslintrc
@@ -5,12 +5,19 @@
         "es6": true,
         "node": true
     },
+    "extends": [
+        "plugin:@typescript-eslint/recommended"
+    ],
+    "parser": "@typescript-eslint/parser",
     "parserOptions": {
         "ecmaFeatures": {
             "jsx": true
         },
         "sourceType": "module"
     },
+    "plugins": [
+        "@typescript-eslint"
+    ],
     "rules": {
         "no-const-assign": "warn",
         "no-this-before-super": "warn",
