+++ .gitignore
@@ -1,4 +1,7 @@
 .idea/
+index.js
+index.js.map
+index.d.ts
 
 # NodeJs
