+++ .npmignore
@@ -1,0 +1,2 @@
+circle.yml
+tsconfig.json
