+++ package.json
@@ -3,9 +3,11 @@
   "version": "1.0.1",
   "description": "Port of Python vocativ library for Czech language",
   "main": "index.js",
+  "types": "index.d.ts",
   "scripts": {
-    "test": "eslint . && mocha",
-    "prepublishOnly": "npm test"
+    "test": "npm run .compile && eslint --ext .js,.ts . && mocha",
+    "prepublishOnly": "npm test",
+    ".compile": "tsc --project ."
   },
   "author": "David Menger",
   "repository": {
@@ -15,8 +17,11 @@
   "license": "MIT",
   "dependencies": {},
   "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^5.59.0",
+    "@typescript-eslint/parser": "^5.59.0",
     "eslint": "^8.38.0",
     "mocha": "^10.2.0",
-    "pickle": "^0.2.0"
+    "pickle": "^0.2.0",
+    "typescript": "^5.0.4"
   }
 }
