+++ test/.eslintrc
@@ -19,6 +19,7 @@
         "no-unreachable": "warn",
         "no-unused-vars": "warn",
         "constructor-super": "warn",
-        "valid-typeof": "warn"
+        "valid-typeof": "warn",
+        "@typescript-eslint/no-var-requires": "off"
     }
 }
