+++ tsconfig.json
@@ -1,0 +1,22 @@
+{
+  "compilerOptions": {
+    "target": "es2015",
+
+    "module": "commonjs",
+    "rootDir": "./",
+    "moduleResolution": "node16",
+    "resolveJsonModule": true,
+
+    "declaration": true,
+    "sourceMap": true,
+    "outDir": "./",
+
+    "esModuleInterop": true,
+    "forceConsistentCasingInFileNames": true,
+
+    "strict": true,
+
+    "skipLibCheck": true
+  },
+  "exclude": []
+}
